separate fetchFunc
